Add basic Spring boot application
Followed one of the spring boot starter tutorials. It was initially initiating alocal repository on startup of the application. I tweaked it to talk to a postgres database running locally. Assuming you have a postgres running locally with a table called employees in the database and a user created in that database for this application then the application would startup successfully.